l4: Fix utcb definition.
The utcb header did break compilation when included.
Reduce types.h a bit and fix utcb.h.

* include/types.h: Don't rely on stdint.h
* include/l4/utcb.h: Include stdint.h, replace l4_thread_t with
  L4_thread_t.

+++ include/types.h
@@ -1,12 +1,8 @@
 #ifndef BRADYPOS_TYPES_h
 #define BRADYPOS_TYPES_H
 
-#include <stdbool.h>
-#include <stddef.h>
-#include <stdint.h>
-
-typedef uint32_t ptr_t;
-typedef uintptr_t memptr_t;
-typedef uint32_t L4_thread_t;
+typedef unsigned ptr_t;
+typedef unsigned memptr_t;
+typedef unsigned L4_thread_t;
 
 #endif
